requirements.txt
torch
cupy-cuda11x
matplotlib
pandas
